update the app

+++ app.py
@@ -9,7 +9,7 @@ from selenium.webdriver.chrome.options import Options
 from requests.exceptions import RequestException, JSONDecodeError
 from json.decoder import JSONDecodeError
 import hashlib,time,os,requests,re
-
+# from selenium.webdriver.firefox.firefox_binary import FirefoxBinary
 
 import os, requests
 
@@ -17,27 +17,38 @@ base_url = "https://www.classcentral.com/"
 base_dir = "Files/"
 
 visited_links = set()
-maxDepth = 1
-
+max_Depth = 1
+1
 def download_page(url, depth=0):
     if url in visited_links:
         return
     visited_links.add(url)
 
-
-    driver = webdriver.Chrome(options=chrome_options)
     driver.get(url)
+    nav_bar = driver.find_element(By.CSS_SELECTOR,"nav")
+    nav_links = nav_bar.find_elements(By.TAG_NAME,"a")
+    links_to_download = []
+
+  
     height = driver.execute_script("return document.body.scrollHeight")
 
-
     html = driver.find_element(By.TAG_NAME,"html")
+    while True:
+        driver.execute_script("window.scrollBy(0, 350)")
+        time.sleep(0.02)
+        newHeight = driver.execute_script("return document.body.scrollHeight")
+        if newHeight == height:
+            break
+        height = newHeight
+    title_element = driver.find_element(By.TAG_NAME, "title")
     try:
 
         for i in range(7):
             body = driver.find_element(By.TAG_NAME,"body")
             body.send_keys(Keys.COMMAND, Keys.SUBTRACT)
         page_dir = base_dir + url.replace(base_url, "").split("/")[0]
-        os.makedirs(page_dir, exist_ok=True)
+        if not os.path.exists(page_dir):
+            os.makedirs(page_dir, exist_ok=True)
         img_elements = driver.find_elements(By.TAG_NAME, "img")
         if len(img_elements) ==0:
             print("No images found on page: ", url)
@@ -56,35 +67,37 @@ def download_page(url, depth=0):
                             if newHeight == height:
                                 break
                             height = newHeight
-                        with open(img_path, "wb") as f:
-                            f.write(requests.get(img_url).content)
-                            time.sleep(0.03)
+                        img_hash = hashlib.md5(img_path.encode()).hexdigest()
+                        if not os.path.exists(img_path):
+                            with open(img_hash, "wb") as f:
+                                f.write(requests.get(img_url).content)
+                                time.sleep(0.03)
                         try:
                             img = WebDriverWait(driver, 10).until(EC.presence_of_element_located((By.TAG_NAME, "img")))
                             driver.execute_script("arguments[0].setAttribute('src', arguments[1]);", img, "file://" + img_path)
                         except StaleElementReferenceException:
                             img = WebDriverWait(driver, 10).until(EC.presence_of_element_located((By.TAG_NAME, "img")))
                             driver.execute_script("arguments[0].setAttribute('src', arguments[1]);", img, "file://" + img_path)
+
                 except JSONDecodeError:
-                    print("JSONDecodeError")
-                    continue
+                    pass
     except RequestException:
-        print("RequestException")
-        continue
+        pass
     except JSONDecodeError:
-        print("JSONDecodeError")
-        continue
+        pass
 
 
     for css in driver.find_elements(By.XPATH, "//link[@rel='stylesheet']"):
         css_url = css.get_attribute("href")
         try:
             if css_url and "http" in css_url:
+                
                 css_path = page_dir + "/" + css_url.split("/")[-1]
-                with open(css_path, "w") as f:
-                    f.write(requests.get(css_url).text)
-                # css.execute_script("arguments[0].setAttribute('value',arguments[1])","href", "file://" + css_path)
-                # css.send_keys("file://" + css_path)
+                css_hash = hashlib.md5(css_path.encode()).hexdigest()
+                if not os.path.exists(css_hash):
+                    with open(css_path, "w") as f:
+                        f.write(requests.get(css_url).text)
+
                 driver.execute_script("arguments[0].setAttribute('href', arguments[1]);", css, "file://" + css_path)
         except JSONDecodeError:
             pass
@@ -94,30 +107,31 @@ def download_page(url, depth=0):
         js_url = js.get_attribute("src")
         try:
             if js_url and "http" in js_url:
-                # js_path = page_dir + "/" + js_url.split("/")[-1]
+
                 file_contents = requests.get(js_url).text
                 file_hash = hashlib.md5(file_contents.encode()).hexdigest()
                 js_path = page_dir + "/" + file_hash + ".js"
-                with open(js_path, "w") as f:
-                    f.write(file_contents)
-                # js.execute_script("arguments[0].setAttribute('value',arguments[1])","src", "file://" + js_path)
-                # js.send_keys("file://" + js_path)
+                if os.path.exists(js_path):
+                    with open(js_path, "w") as f:
+                        f.write(file_contents)
+    
                 driver.execute_script("arguments[0].setAttribute('src', arguments[1]);", js, "file://" + js_path)
         except JSONDecodeError:
             pass
 
+    
     with open(page_dir + "/index.html", "w") as f:
         f.write(driver.page_source)
-    if depth < maxDepth:
+    driver.implicitly_wait(1)
+    if depth < max_Depth:
         for link in driver.find_elements(By.TAG_NAME, "a"):
             link_url = link.get_attribute("href")
             if link_url and "http" in link_url and base_url in link_url:
-                if link_url.endswith("/"):
-                    download_page(link_url)
-                else:
-                    download_page(link_url + "/")
-# binary = FirefoxBinary('./geckodriver')
-driver = webdriver.Chrome()
+                if link_url.endswith('/'):
+                    download_page(link_url, depth+1)
+            else:
+                download_page(link_url + '/',depth+1)
+driver = webdriver.Firefox()
 
 driver.get(base_url)
